Move `GetSettingUnderlyingValue` to a `SettingSource` extension method
